introduce P control to both motors

+++ Bontaco/bontaco_drive.h
@@ -1,7 +1,8 @@
 #include "iodefine.h"
 #include "bontaco_common.h"
+#include "bontaco_encoder.h"
 
 void run_straight();
-void set_target_speed(unsigned int);
+void set_target_speed(float);
 void pid_control_motor_left();
 void pid_control_motor_right();
